+++ Traktor.Core/Domain/Trakt/Show.cs
@@ -38,7 +38,7 @@ namespace Traktor.Core.Domain.Trakt
         public object trailer { get; set; }
         public string homepage { get; set; }
         public string status { get; set; }
-        public int rating { get; set; }
+        public decimal rating { get; set; }
         public int votes { get; set; }
         public int comment_count { get; set; }
         public string language { get; set; }
